refactor: diagnosis, main 컴포넌트 분리
