+++ src/pages/diagnosis/answerButtons/index.tsx
@@ -1,68 +1,10 @@
-import styled from "styled-components";
-import { IAnswerButtonProps } from "../../interfaces/diagnosisPage";
-import RoundButton from "../roundButton";
-import theme from "../../lib/theme";
-import { Body_1 } from "../../lib/fontStyle";
+import { IAnswerButtonProps } from "../../../interfaces/diagnosisPage";
+import RoundButton from "../../../components/roundButton";
+import theme from "../../../lib/theme";
 import { useEffect } from "react";
-import { saveAnswer } from "../../state/answerSlice";
-import { useAppDispatch } from "../../state";
-
-const Container = styled.section`
-  background: transparent;
-  width: 100%;
-
-  margin-bottom: 13rem;
-`;
-const AnswersContainer = styled.section<{ ansCount: number }>`
-  width: 100%;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  margin-top: ${({ ansCount }) => (ansCount === 2 ? 13.4 : ansCount === 3 ? 10.2 : 7)}rem;
-`;
-const ButtonBox = styled.section`
-  width: calc(var(--vw, 1vw) * 100 - 4rem);
-  & + & {
-    margin-top: 1.2rem;
-  }
-`;
-const Button = styled.section<{ selected: boolean }>`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  padding: ${({ selected }) => (selected ? "1.4rem 2.3rem" : "1.3rem 2.3rem")};
-
-  background-color: ${({ selected, theme }) => (selected ? theme.color.sub_blue : "transparent")};
-  color: ${({ selected, theme }) => (selected ? theme.color.blue : theme.color.grey_300)};
-
-  border: ${({ selected, theme }) => !selected && `0.1rem solid ${theme.color.grey_650}`};
-  border-radius: 9rem;
-
-  cursor: pointer;
-`;
-const ButtonText = styled(Body_1)`
-  font-weight: 200;
-  text-align: center;
-
-  width: 26rem;
-`;
-const NextButton = styled.section`
-  width: calc(var(--vw, 1vw) * 100);
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  position: fixed;
-  bottom: 0;
-  padding-bottom: 3rem;
-  padding-top: 0.6rem;
-
-  background: linear-gradient(180deg, rgba(31, 37, 79, 0) 0%, #23284b 50%);
-`;
+import { saveAnswer } from "../../../state/answerSlice";
+import { useAppDispatch } from "../../../state";
+import { Container, AnswersContainer, ButtonBox, ButtonText, NextButton } from "./index.style";
 
 const AnswerButtons = ({ question, selectedAnswer, setSelectedAnswer, handleNext }: IAnswerButtonProps) => {
   const answers = question.answers;
@@ -122,10 +64,10 @@ const AnswerButtons = ({ question, selectedAnswer, setSelectedAnswer, handleNext
       <AnswersContainer ansCount={answers.length}>
         {answers.length !== 0 &&
           answers.map((ans, idx) => (
-            <ButtonBox key={idx} onClick={() => handleSelect(ans.answer_id)}>
-              <Button selected={handleActive(ans.answer_id)}>
+            <ButtonBox key={idx} onClick={() => handleSelect(ans.answer_id)} selected={handleActive(ans.answer_id)}>
+              <section className="button">
                 <ButtonText>{ans.answer}</ButtonText>
-              </Button>
+              </section>
             </ButtonBox>
           ))}
       </AnswersContainer>
